Add files via upload
Function predicts a cost reduced flux distribution at a pre-defined percentage from optima. Here, a cost is first associated with each reaction in the network. Next, the objective function is fixed at a predetermined value, and this cost is optimized throughout the model, thus predicting a cost-optimal flux distribution. This function was used in the original publication, Schultz and Qutub, 2015, BMC systems biology, 9(1), 18, to predict internal cell fluxes in E. coli.